Change look of current ride page and connect to back

+++ rocket_back/rocket/src/main/java/com/uber/rocket/entity/ride/Ride.java
@@ -47,10 +47,10 @@ public class Ride {
     private Boolean splitFare;
 
     @NotNull
-    private int price;
+    private double price;
 
     @NotNull
-    private int duration;
+    private double duration;
 
     private double length;
